// src/models/playlist.model.ts
import {
  FETCH_PLAYLIST_FAILURE,
  FETCH_PLAYLIST_REQUEST,
  FETCH_PLAYLIST_SUCCESS,
} from "../redux/types/playlistTypes";

export interface RootPlaylists {
  collaborative: boolean
  description: string
  external_urls: string
  followers: Followers
  href: string
  id: string
  images: Image[]
  name: string
  owner: Owner
  primary_color: string
  public: boolean
  snapshot_id: string
  tracks: Tracks
  type: string
  uri: string
}

export interface ExternalUrls {
  spotify: string
}

export interface Followers {
  href: any
  total: number
}

export interface Image {
  height: any
  url: string
  width: any
}

export interface Owner {
  display_name: string
  external_urls: string
  href: string
  id: string
  type: string
  uri: string
}



export interface Tracks {
  href: string
  items: Item[]
  limit: number
  next: any
  offset: number
  previous: any
  total: number
}

export interface Item {
  added_at: string
  added_by: AddedBy
  is_local: boolean
  primary_color?: string
  track: Track
  video_thumbnail: VideoThumbnail
}

export interface AddedBy {
  external_urls: ExternalUrls3
  href: string
  id: string
  type: string
  uri: string
}

export interface ExternalUrls3 {
  spotify: string
}

export interface Track {
  album: Album
  artists: Artist2[]
  available_markets: string[]
  disc_number: number
  duration_ms: number
  episode: boolean
  explicit: boolean
  external_ids: ExternalIds
  external_urls: ExternalUrls7
  href: string
  id: string
  is_local: boolean
  name: string
  popularity: number
  preview_url?: string
  track: boolean
  track_number: number
  type: string
  uri: string
}

export interface Album {
  album_type: string
  artists: Artist[]
  available_markets: string[]
  external_urls: ExternalUrls5
  href: string
  id: string
  images: Image2[]
  name: string
  release_date: string
  release_date_precision: string
  total_tracks: number
  type: string
  uri: string
}

export interface Artist {
  external_urls: ExternalUrls4
  href: string
  id: string
  name: string
  type: string
  uri: string
}

export interface ExternalUrls4 {
  spotify: string
}

export interface ExternalUrls5 {
  spotify: string
}

export interface Image2 {
  height: number
  url: string
  width: number
}

export interface Artist2 {
  external_urls: ExternalUrls6
  href: string
  id: string
  name: string
  type: string
  uri: string
}

export interface ExternalUrls6 {
  spotify: string
}

export interface ExternalIds {
  isrc: string
}

export interface ExternalUrls7 {
  spotify: string
}

export interface VideoThumbnail {
  url: any
}


export interface PlaylistState {
  data: RootPlaylists;
  loading: boolean;
  error: string | null;
}

export interface FetchPlaylistSuccessPayload {
  data: RootPlaylists;
}

export interface FetchPlaylistFailurePayload {
  error: string;
}

export interface FetchPlaylistRequest {
  type: typeof FETCH_PLAYLIST_REQUEST;
  id: string;
}

export type FetchPlaylistSuccess = {
  type: typeof FETCH_PLAYLIST_SUCCESS;
  payload: FetchPlaylistSuccessPayload;
};

export type FetchPlaylistFailure = {
  type: typeof FETCH_PLAYLIST_FAILURE;
  payload: FetchPlaylistFailurePayload;
};

export type PlaylistActions =
  | FetchPlaylistRequest
  | FetchPlaylistSuccess
  | FetchPlaylistFailure;